requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiolimiter==1.2.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blobfile==3.0.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.4
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
einops==0.8.1
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastavro==1.11.1
fastembed==0.6.1
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
git-filter-repo==2.47.0
greenlet==3.2.3
griffe==1.7.3
groq==0.25.0
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.0
llama-index-cli==0.5.0
llama-index-core==0.13.0
llama-index-embeddings-huggingface==0.6.0
llama-index-embeddings-openai~=0.5.0
llama-index-embeddings-voyageai==0.4.1
llama-index-indices-managed-llama-cloud==0.9.1
llama-index-instrumentation==0.2.0
llama-index-llms-groq==0.4.0
llama-index-llms-mistralai==0.7.0
llama-index-llms-ollama==0.7.0
llama-index-llms-openai
llama-index-llms-openai-like==0.5.0
llama-index-llms-together==0.4.0
llama-index-multi-modal-llms-openai==0.6.0
llama-index-postprocessor-cohere-rerank==0.5.0
llama-index-readers-file==0.5.0
llama-index-readers-github==0.8.0
llama-index-readers-llama-parse==0.5.0
llama-index-storage-chat-store-redis==0.5.0
llama-index-storage-docstore-redis==0.4.0
llama-index-storage-index-store-redis==0.5.0
llama-index-storage-kvstore-redis==0.4.0
llama-index-vector-stores-qdrant==0.7.1
llama-index-workflows==1.0.1
llama-parse==0.6.12
llama_stack==0.2.1
llama_stack_client==0.2.8
loguru==0.7.3
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistralai==1.9.1
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
ollama==0.5.1
onnxruntime==1.22.0
openai==1.82.0
optimum==1.26.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.8
portalocker==2.10.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.8
py_rust_stemmers==0.1.5
pyaml==25.5.0
pyasn1==0.4.8
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyGithub==2.7.0
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pypdf==5.7.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.2
redis==5.3.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
tqdm==4.67.1
transformers==4.53.2
tree-sitter==0.24.0
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.23.2
tree-sitter-language-pack==0.8.0
tree-sitter-yaml==0.7.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
upstash-redis==1.4.0
urllib3==2.5.0
uvicorn==0.34.2
voyageai==0.3.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
